+++ src/main/java/org/patternfly/thirdparty/popper/Rect.java
@@ -20,10 +20,10 @@ import jsinterop.annotations.JsType;
 import static jsinterop.annotations.JsPackage.GLOBAL;
 
 @JsType(isNative = true, namespace = GLOBAL, name = "Object")
-public class State {
+public class Rect {
 
-    public StateElements elements;
-    public String placement;
-    public StateRects rects;
-    public StyleDefinitions styles;
+    public double x;
+    public double y;
+    public double width;
+    public double height;
 }