+++ src/main/java/org/patternfly/thirdparty/popper/StateRects.java
@@ -20,10 +20,8 @@ import jsinterop.annotations.JsType;
 import static jsinterop.annotations.JsPackage.GLOBAL;
 
 @JsType(isNative = true, namespace = GLOBAL, name = "Object")
-public class State {
+public class StateRects {
 
-    public StateElements elements;
-    public String placement;
-    public StateRects rects;
-    public StyleDefinitions styles;
+    public Rect reference;
+    public Rect popper;
 }